fix windows SyncDir issue
