+++ ci/scripts/create_binary_package.sh
@@ -8,4 +8,7 @@ make "release/${TARGET}"
 mkdir "release/${TARGET}/config"
 mv sampleconfig/*yaml "release/${TARGET}/config"
 cd "release/${TARGET}"
+if [ "$TARGET" == "windows-amd64" ]
+    for FILE in bin/*; do mv $FILE $FILE.exe; done
+fi
 tar -czvf "hyperledger-fabric-${TARGET}-${RELEASE}.tar.gz" bin config
